Add command to run app against stubs for development
