+++ src/electron-starter.js
@@ -9,16 +9,6 @@ const setupProtocols = require('./electronHelpers/setupProtocols');
 
 const { app, shell, Menu } = electron;
 
-// Set server port depending on electron environment
-let serverPort = isDev ? '9821' : '9822';
-
-// Override if running electron in dev mode against built production files
-if (process.env.NODE_ENV === 'production') {
-  serverPort = '9822';
-}
-
-require('../server')(serverPort);
-
 require('electron-debug')({ enabled: true });
 
 // Module to create native browser window.
